fewer windows warnings
git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@4351 3d70eeeb-363e-0410-a505-8a46323a89f2

+++ src/limn/polyshapes.c
@@ -534,8 +534,8 @@ limnPolyDataSpiralSuperquadric(limnPolyData *pld,
         yy = (pld->xyzw + 4*vertIdx)[1];
         rr = sqrt(xx*xx + yy*yy);
         if (rr < 0.1 && rr) {
-          (pld->xyzw + 4*vertIdx)[0] *= AIR_CAST(double, 0.1/rr);
-          (pld->xyzw + 4*vertIdx)[1] *= AIR_CAST(double, 0.1/rr);
+          (pld->xyzw + 4*vertIdx)[0] *= AIR_CAST(float, 0.1/rr);
+          (pld->xyzw + 4*vertIdx)[1] *= AIR_CAST(float, 0.1/rr);
         }
       }
       if ((1 << limnPolyDataInfoNorm) & infoBitFlag) {
